Updated Strategies and Factory:
* **SingleEventCopyStrategy:**
    * Format: `copy event <eventName> on <dateStringTtimeString> --target <calendarName> to <dateStringTtimeString>`
    * Improved validation to check each component of the command in sequence.
    * Added specific error messages for each part of the command format.
* **DayEventsCopyStrategy:**
    * Format: `copy events on <dateString> --target <calendarName> to <dateString>`
    * Enhanced validation to ensure the exact command structure.
    * Added specific error messages for each keyword and parameter.
* **RangeEventsCopyStrategy:**
    * Format: `copy events between <dateString> and <dateString> --target <calendarName> to <dateString>`
    * Added thorough validation of the command structure.
    * Improved error handling for the date range parameters.
* **CopyStrategyFactory:**
    * Enhanced the strategy selection logic to better match the exact command formats.
    * Added explicit checks for the command structure rather than relying on the `canHandle` method.
    * Prioritized matching the three specific formats from the assignment.
